Add ClickHouse 25.12 to CI and fix issues
Replace 25.11 with 25.12 and replace 25.3.9 with 25.3.10. The pattern
here is to run the latest release plus all previous lts releases.

Oddly, 25.12 returns an error string with a trailing newline on auth
failure. So strip it out.

Also, run `make indent` and let it change the indentation of a query in
`chfdw_construct_create_tables()`.