Remove the custom code for Adjust extensions
There was a whole lot of code to support a number of Ajust GmbH
extensions inherited from the old clickhouse_fdw code: [ajbool], ajtime,
[country], and [istore]. Presumably they would allow those extensions'
custom data types and functions to work with ClickHouse.

pg_clickhouse takes a different approach, aiming to support all the data
types in ClickHouse and PostgreSQL, but not non-core custom PostgreSQL
data type extensions unless they prove essential to map to specific
ClickHouse types.

So remove all the code for those extensions.

Leave, however, the code that supports the in-core intarray and hstore
extensions. Expect future work to flesh out that support.

  [ajbool]: https://pgxn.org/dist/ajbool/ "ajbool on PGXN"
  [country]: https://pgxn.org/dist/country/ "country on PGXN"
  [istore]: https://pgxn.org/dist/istore/ "istore on PGXN"